+++ tests/base_tests/test_sqlquery_info.py
@@ -100,7 +100,7 @@ async def test_very_simple_condition():
 
     for i in SQL_OP.ALL:
         sqi = SQLQueryInfo()
-        if i in SQL_OP.IN.value:
+        if i in SQL_OP.IN.value or i in SQL_OP.CONTAINS.value or i in SQL_OP.CONTAINS_ANY.value:
             sqi.parse_then_add_condition('a', i, '[1,2]')
             assert sqi.conditions[0] == ['a', SQL_OP.txt2op[i], [1,2]]
         else:
@@ -112,7 +112,7 @@ class ATestModel(Model):
     name = TextField()
 
 
-@app.route('test1')
+@app.route.view('test1')
 class ATestView(PeeweeView):
     model = ATestModel