+++ tests/support_tests/test_peewee.py
@@ -24,7 +24,7 @@ class ATestModel(Model):
         database = db
 
 
-@app.route('test')
+@app.route.view('test')
 class ATestView(PeeweeView):
     model = ATestModel